A2DN_userhtml_v6 controller refactoring and login page customisation
